travis: unset GOROOT
GOROOT is set for preinstalled go compiler v1.4.
And on go-getting it tries to search for library under the wrong GOROOT.
Since we manually install golang into /usr/local/go which is assumed by
the go system as the default GOROOT. Then we should unset the GOROOT
before everything so go system can access the correct search path.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>